Add files via upload
Action to monitoring/ audit agents from BPO coworkers, in parallel to current systems, this kind of analysis consists in  quantify the biggest quantity of interaction in less time. The sum of point can measure, filtering by date or category, moreover by the agent,. Is possible add any new paramater, and generate graphs and PBI bases with this source.